Sort steps by timestamps in a TaskRun
Currently, all injected steps appear before the regularly defined
steps in some views. Now they are sorted by the finishedAt timestamp
in each step's status, with ties broken by the startedAt timestamp.

The current order is not always correct, since some injected steps
actually run after regular steps. This ordering also causes some steps
to be incorrectly labelled as either skipped or failed.

The new order should be better, but it will still depend on the
precision of the timestamps: seconds might not be sufficient in all
cases. There are plans in 'pipeline' to increase the precision.

Fixes #1367